Relocate, alphabetize & add comments to TagHelperCatalog: - Moved services/interface  ITagHelperService previously created to proper location in existing interfaces folder & adjusted namespace - TagHelperService: Cleanup  - Alphabetize the properties & functions - TagHelperService: started listing each helper’s bindable public properties (fixed missed properties at load time) - Added tagHelper reference View models - Added proper comments to TagHelperCatalogService
