+++ GCFoundation.Web/Models/Components/ComponentsIndexPageViewModel.cs
@@ -1,3 +1,5 @@
+using GCFoundation.Components.Models;
+
 namespace GCFoundation.Web.Models.Components
 {
     /// <summary>
